UPSTREAM: amd/hudson/pi: Fix position of hudson_fwm
AMDFWTOOL calculates the location of the amdfw based on the
CONFIG_ROM_SIZE. If HUDSON_FWM_POSITION does not match that location the
resulting rom does not boot.

This patch forces the position of HUDSON_FWM_POSITION to be the
position calculated by amdfwrom.

Tested on a Bettong derivative with a 16MiB flash.

BUG=none
BRANCH=none
TEST=none

Change-Id: Iddd79a218e41611adda77c5e52366e8857d8b084
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: baae959a639712ac34b74b307b9fad5cc8b2c663
Original-Change-Id: Id2ee96ee076293d48ade84fd6e976ca994dcf491
Original-Signed-off-by: Ricardo Ribalda Delgado <ricardo.ribalda@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/17925
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/427471