UPSTREAM: amd/pi: Make BottomIo position configurable
Some PCI peripherals, such as FPGA accelerators, require a great amount
of memory mapped IO. This patch allows the user to select at build time
the bottom IO to leave enough space for such devices.

We cannot calculate this value at runtime because it has to be set
before the PCI devices are enumerated.

BUG=none
BRANCH=none
TEST=none

Change-Id: If3289ef28c08a2e18b0095ff2be3091e40f1acfa
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: a132892de696d011e61cea2828c0e4c93fc3b3aa
Original-Change-Id: Ic590e8aa8b91ff89877cbff6afd10614d33dcf8d
Original-Credit-to: Kysti Mlkki <kyosti.malkki@gmail.com>
Original-Signed-off-by: Ricardo Ribalda Delgado <ricardo.ribalda@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/17980
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/427463