UPSTREAM: sb/intel/fsp_rangeley: Fix NULL check in gpio.c
This should always have been an and, not an or.

The only way this would happen is if no GPIOs were getting configured,
so we shouldn't ever have a NULL here, but if we did, GPIOs would
be randomly configured, which would have 'interesting' results.

BUG=none
BRANCH=none
TEST=none

Change-Id: I480ac3acde34c2f3d44424a001065f37a2428d72
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: 6e4cb50420b7ddfecf9388e8a3b000ecb6855054
Original-Found-by: Coverity Scan #1229633 & 1229632
Original-Change-Id: If123372658383f84279738e1186425beba3208ca
Original-Signed-off-by: Martin Roth <martinroth@google.com>
Original-Reviewed-on: https://review.coreboot.org/18095
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/428255
Commit-Ready: Aaron Durbin <adurbin@chromium.org>