UPSTREAM: Only add etc/ps2-keyboard-spinup when not updating an image
Without this motherboards that requires a non zero timeout for ps2
keyboards on SeaBIOS don't build when CONFIG_UPDATE_IMAGE is set.

An alternative way to achieve this file would be to include a cbfsfile
instead of calling cbfstool. That way the file gets updated/added both
both image update and regular build. A difficulty of that approach is
that it needs to convert a decimal to a binary in little endian
representation, which is not a trivial thing to do in a Makefile.

BUG=none
BRANCH=none
TEST=none

Change-Id: I4814675b2f006656d1bd6f418e38afadee6aeb80
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: 7218a1e754d84f84fd13bd5a66b38c6f0eb6f884
Original-Change-Id: Icafba8d3e279a2e70e607abba81e3dbebfb55e4b
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/18231
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/433883