UPSTREAM: mainboard/intel: Add MinnowBoard 3
This commit adds the initial scaffolding for the MinnowBoard 3
with Apollo Lake silicon.

This mainboard is based on Intel's Leafhill CRB with Apollo Lake
silicon. In a first step, it concerns only a copy of intel/leafhill
directory with name changes. Special adaptations for MinnowBoard 3
mainboard will follow in separate commits.

BUG=none
BRANCH=none
TEST=none

Change-Id: I266e9f12db1c4824545871e6a0d1ac89f8d8255f
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: 35f03d902700e5a9d99a561516588169598b0f07
Original-Change-Id: I7563fe37c89511c7035c5bffc9b034b379cfcaf4
Original-Signed-off-by: Brenton Dong <brenton.m.dong@intel.com>
Original-Reviewed-on: https://review.coreboot.org/18298
Original-Tested-by: build bot (Jenkins)
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/445834