UPSTREAM: ec/ene_kb3940q: correct ACPI battery data fed into ToString()
ToString() requires the input buffer data to be null-terminated,
but the data returned by the EC is not, leading Windows to fail
to report any battery data at all.  Correct this by concatenating
a null terminator (0x00) to the end of the buffer data before
inputting to ToString() where needed

TEST: boot Windows on google/butterfly, observe battery data
reported correctly.

BUG=none
BRANCH=none
TEST=none

Change-Id: I4b79699fd993574addaf1a8cc0dfdb43e1f575b5
Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Commit-Id: af4c0a431c24a38519815a6e44325d29f9073da7
Original-Change-Id: I395cc7fbdf26c8cc816e47107e552c0533580fa1
Original-Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/19961
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/523979